- Add validation and controller advice if no company is found in delete and add test

+++ src/main/java/com/thoughtworks/parking_lot/Service/ParkingLotService.java
@@ -3,8 +3,11 @@ package com.thoughtworks.parking_lot.Service;
 import com.thoughtworks.parking_lot.Entity.ParkingLot;
 import com.thoughtworks.parking_lot.Repository.ParkingLotRepo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.PathVariable;
+
+import java.util.Objects;
 
 @Service
 public class ParkingLotService {
@@ -18,7 +21,10 @@ public class ParkingLotService {
 
     public ParkingLot delete(String name) {
         ParkingLot fetchedParkingLot = parkingLotRepo.findByName(name);
-        parkingLotRepo.delete(fetchedParkingLot);
-        return fetchedParkingLot;
+        if(Objects.nonNull(fetchedParkingLot)){
+            parkingLotRepo.delete(fetchedParkingLot);
+            return fetchedParkingLot;
+        }
+        return null;
     }
 }
